move light stuff out of decal cluster
