+++ crates/bevy_pbr/src/light/mod.rs
@@ -26,11 +26,11 @@ mod ambient_light;
 pub use ambient_light::AmbientLight;
 
 mod point_light;
-pub use point_light::PointLight;
+pub use point_light::{PointLight, PointLightTexture};
 mod spot_light;
-pub use spot_light::SpotLight;
+pub use spot_light::{SpotLight, SpotLightTexture};
 mod directional_light;
-pub use directional_light::DirectionalLight;
+pub use directional_light::{DirectionalLight, DirectionalLightTexture};
 
 /// Constants for operating with the light units: lumens, and lux.
 pub mod light_consts {
